+++ s4e-backend/src/main/java/pl/cyfronet/s4e/ex/product/ProductException.java
@@ -1,0 +1,22 @@
+package pl.cyfronet.s4e.ex.product;
+
+public abstract class ProductException extends Exception {
+    public ProductException() {
+    }
+
+    public ProductException(String message) {
+        super(message);
+    }
+
+    public ProductException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public ProductException(Throwable cause) {
+        super(cause);
+    }
+
+    public ProductException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
+        super(message, cause, enableSuppression, writableStackTrace);
+    }
+}